+++ app/models/concerns/intimacies.rb
@@ -7,5 +7,13 @@ module Intimacies
   included do
     validates :principles, json: { schema: Schemas::INTIMACY }
     validates :ties,       json: { schema: Schemas::INTIMACY }
+
+    def without_secrets
+      c = dup
+      c.id = id
+      c.ties.reject! { |t| t['hidden'] }
+      c.principles.reject! { |p| p['hidden'] }
+      c
+    end
   end
 end